app/requirements.txt
flask==3.0.3
openai==1.76.2
boto3==1.34.131
python-dotenv==1.0.1
pytest==8.3.2
